% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bea_region}
\alias{bea_region}
\title{BEA Region Codes}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 51 rows and 7 columns.}
\source{
\url{https://www.bea.gov/regional/docs/regions.cfm}
}
\usage{
bea_region
}
\description{
Component state list for BEA regions.
}
\keyword{datasets}
