% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{census_region_division}
\alias{census_region_division}
\title{Census Bureau Region and Division Codes}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 51 rows and 7 columns.}
\source{
\href{https://www2.census.gov/programs-surveys/popest/geographies/2015/}{U.S. Census Bureau, Population Division}
}
\usage{
census_region_division
}
\description{
Contains crosswalk of census region and division and state FIPS code.
}
\keyword{datasets}
