% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{county}
\alias{county}
\title{County-level FIPS Code}
\format{A data frame with five variables:
\describe{
\item{usps}{Official United States Postal Service (USPS) Code}
\item{state}{State Name}
\item{fips}{2010 FIPS County Code}
\item{county}{Full County Name}
\item{cty_short}{Short County Name}
}}
\source{
\url{https://www.census.gov/geo/reference/codes/cou.html}
}
\usage{
county
}
\description{
2010 FIPS Codes for Counties and County Equivalent Entities.
}
\keyword{datasets}

