% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{county_ipums_usa}
\alias{county_ipums_usa}
\title{Listing of Counties Identified in IPUMS USA}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 589 rows and 17 columns.}
\source{
\url{https://usa.ipums.org/usa-action/variables/COUNTYFIP#description_section}
}
\usage{
county_ipums_usa
}
\description{
Contains counties that are identified through the microdata maintained by
IPUMS.
}
\keyword{datasets}
