% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sleeptimesformat.R
\name{expand_sleep_series}
\alias{expand_sleep_series}
\title{Expand Sleep Times to full vector}
\usage{
expand_sleep_series(.stdata, expand_by = 5)
}
\arguments{
\item{.stdata}{A sleeptimes dataframe}

\item{expand_by}{Amount (in minutes) to expand sleep times by}
}
\value{
Sleeptimedataframe with single columns vector for datetime and wake status
}
\description{
Turns the paired sleeptimes into a long single vectored datetime sequence
}
\keyword{internal}
