% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fips}
\alias{fips}
\title{FIPS data}
\format{A data frame with three variables: #' \describe{
\item{fips}{FIPS State Code}
\item{usps}{Official United States Postal Service (USPS) Code}
\item{state}{State Name}
}}
\source{
\url{https://www.census.gov/geo/reference/ansi_statetables.html}
}
\usage{
fips
}
\description{
FIPS Codes for the States and the District of Columbia form Census.
}
\keyword{datasets}

