% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIPS_plot.R
\name{FIPS_collect}
\alias{FIPS_collect}
\title{Build a collection of FIPS simulations for overlay plotting}
\usage{
FIPS_collect(...)
}
\arguments{
\item{...}{either named FIPS_simulation objects, or a single named list of them}
}
\value{
an object of class FIPS_collection
}
\description{
Build a collection of FIPS simulations for overlay plotting
}
