% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIPS_df.R
\name{FIPS_df}
\alias{FIPS_df}
\title{The FIPS_df}
\usage{
FIPS_df(.data)
}
\description{
The only input \emph{required} to generate this is a series of sleep/wake times.
The FIPS_df is a tibble with the following variables (columns):

\code{datetime} = vector of datetime stamps separated at equidistance intervals.
\code{sleep.id} = a supplementary variable indicating sleep episode identifier.
\code{wake_status} =  Awake (\code{T}) or asleep (\code{F}) at that epoch interval/epoch
\code{wake_status_int} = Awake (1) or asleep (0) at that epoch interval/epoch
\code{change_point} = Whether the individual changed wake status at that interval/epoch.
\code{switch_direction} = Whether switch was to sleep or to wake
\code{status_duration} = How long individual has been in status at that current time point
\code{total_prev} = If a switch has occured, how long were that in the previous status.
\code{time} = time of day in decimal hour
\code{day} =  days into simulation
\code{sim_hours} = hours simulation has run for total
}
\details{
All models implemented in FIPS are implemented to be run on a \code{FIPS_df} ---
a dataframe containing a time series of all variables required to run \link{FIPS_simulate} to
generate a FIPS_simulation object.
}
\seealso{
See FIPS_simulation (internal) for additional columns added after a simulation is run.

Also see \link{parse_sleeptimes} for sleep times converter.
}
