% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIPS_plot.R
\name{FIPS_plot}
\alias{FIPS_plot}
\title{FIPS Time Series Plot}
\usage{
FIPS_plot(
  dats,
  from = NULL,
  to = NULL,
  plot_stat = NULL,
  fatigue_CIs = FALSE,
  observed = NULL,
  observed_y = NULL
)
}
\arguments{
\item{dats}{A FIPS_simulation object (i.e., FIPS_df with simulation results)}

\item{from}{The starting datetime to be plotted}

\item{to}{The ending datetime to be plotted}

\item{plot_stat}{Which variables to plot}

\item{fatigue_CIs}{A logical indicating whether uncertainty intervals on fatigue should be plotted}

\item{observed}{A data frame with any observed sleepiness ratings or objective indicators to plot against predictions}

\item{observed_y}{The name of the observed sleepiness ratings in the observed data frame}
}
\value{
A ggplot2 object displaying fatigue and other requested processes over time
}
\description{
FIPS Time Series Plot
}
