% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sequencefunctions.R
\name{generate_decimal_timeunits}
\alias{generate_decimal_timeunits}
\title{Generate Decimal Time dates from a timedate object}
\usage{
generate_decimal_timeunits(.data, datetime.col)
}
\description{
Requires a datetime.col, will return the time day and hours since first timedate
all in decimal time
}
