% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sleeptimesformat.R
\name{generate_presleep_times}
\alias{generate_presleep_times}
\title{Fill pre-observation wake times}
\usage{
generate_presleep_times(simulationstart, firstsleep, expand_by = 5)
}
\arguments{
\item{simulationstart}{start of simulation}

\item{firstsleep}{first sleep in the sleep dataframe}

\item{expand_by}{expand}
}
\value{
returns expanded tibble containing sleep.id = NA (due to waking) and wake_status = T
}
\description{
The first sleep is unlikely to also be the start of the mission simulation
Thus, this function fills the start of the tibble with the all times between
The mission start time and the first instance of sleep, intervaled by X minutes
}
