% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIPS_formula_interface.R
\name{get_bmm_model_frame}
\alias{get_bmm_model_frame}
\title{get_bmm_model_frame}
\usage{
get_bmm_model_frame(.FIPS_sim, model_formula)
}
\arguments{
\item{.FIPS_sim}{A FIPS_Simulation object}

\item{model_formula}{A formula describing how the time-varying processes predictors should be calculated for the predicted output.}
}
\value{
Minimal dataframe with only requird columns for computation
}
\description{
get_bmm_model_frame
}
