% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIPS_df.R
\name{is_FIPS_df}
\alias{is_FIPS_df}
\title{Test if the object is a \link{FIPS_df}}
\usage{
is_FIPS_df(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} if the object inherits from the \code{inherits(x, "FIPS_df")}.
}
\description{
This function returns \code{TRUE} for FIPS_df,
and \code{FALSE} for all other objects, including regular data frames.
}
