% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIPS_simulation.R
\name{is_FIPS_simulation}
\alias{is_FIPS_simulation}
\title{Test if the object is a simmed FIPS_df}
\usage{
is_FIPS_simulation(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} if the object inherits from the \code{inherits(x, "FIPS_df") & attr(x, "simmed") }.
}
\description{
This function returns \code{TRUE} for FIPS_df if a simulation has been run on it,
and \code{FALSE} for all other objects, including regular data frames.
}
