% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lower48}
\alias{lower48}
\title{State-level FIPS Code}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 49 rows and 3 columns.}
\source{
\url{https://www.census.gov/geo/reference/ansi_statetables.html}
}
\usage{
lower48
}
\description{
FIPS Codes for the 48 Contiguous Continental States and the District of Columbia.
}
\keyword{datasets}
