% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_mccauley2013.R
\docType{data}
\name{mccauley2013_pvec}
\alias{mccauley2013_pvec}
\title{Default parameter vector for the Differential Performance Model}
\format{
An object of class \code{numeric} of length 17.
}
\usage{
mccauley2013_pvec
}
\description{
Default parameter vector for the Differential Performance Model
}
\keyword{datasets}
