% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_mccauley2013.R
\name{mccauley2013_simulate}
\alias{mccauley2013_simulate}
\title{Simulate: McCauley Model}
\usage{
mccauley2013_simulate(pvec, dat, substep_minutes = 2)
}
\arguments{
\item{pvec}{Parameter vector, see [mccauley2013_pvec]}

\item{dat}{Input dataframe (ensure this is a FIPS_df)}
}
\value{
Dataframe with simulation results
}
\description{
Runs the McCauley et al. (2013) PVT model over the supplied FIPS_df.
}
