% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_mccauley2013.R
\name{mccauley2013_simulation_dispatch}
\alias{mccauley2013_simulation_dispatch}
\title{McCauley Simulation Dispatcher}
\usage{
mccauley2013_simulation_dispatch(
  dat,
  pvec,
  model_formula = NULL,
  substep_minutes = 2
)
}
\arguments{
\item{dat}{A FIPS_df}

\item{pvec}{Parameter vector}

\item{model_formula}{Optional formula for custom predictions}
}
\description{
McCauley Simulation Dispatcher
}
