% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_mccauley2024.R
\name{mccauley2024_check_pvec}
\alias{mccauley2024_check_pvec}
\title{Check McCauley Parameter Vector}
\usage{
mccauley2024_check_pvec(pvec)
}
\arguments{
\item{pvec}{The pvec to check contains all required three process model parameters}
}
\value{
logical
}
\description{
Checks the pvec contains required parameters.
}
