% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_mccauley2024.R
\docType{data}
\name{mccauley2024_pvec}
\alias{mccauley2024_pvec}
\title{Default parameter vector for the Differential Performance Model}
\format{
An object of class \code{numeric} of length 25.
}
\usage{
mccauley2024_pvec
}
\description{
Default parameter vector for the Differential Performance Model
}
\keyword{datasets}
