% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nchs_urc}
\alias{nchs_urc}
\title{NCHS Urban-Rural Codes}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3147 rows and 10 columns.}
\source{
\url{https://www.cdc.gov/nchs/data_access/urban_rural.htm}
}
\usage{
nchs_urc
}
\description{
2013, 2006, and 1990 NCHS Urban-Rural Classification Scheme for Counties.
}
\keyword{datasets}
