% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_singlevectorformat.R
\name{parse_sleepwake_sequence}
\alias{parse_sleepwake_sequence}
\title{Parse Sleep Wake Binary/Integer Sequence}
\usage{
parse_sleepwake_sequence(seq, epoch, series.start)
}
\arguments{
\item{seq}{A sequence of \code{0} (sleep) and \code{1} (wake) integers indicating sleep/wake status at that moment.}

\item{epoch}{Integer expressing length of each observations in the series (minutes).}

\item{series.start}{A POSIXct object indicating the start datetime of the simulation (i.e., pre-first sleep waking duration)}
}
\value{
\code{FIPS_df} formatted dataframe
}
\description{
It is common to have sleep wake history information in the form of a binary sequence.
This is the format used by SAFTE-FAST and other proprietary software.
Further, this format is often easily exported by actigraphy measurement software.
}
