% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIPS_plot.R
\name{plot.FIPS_simulation}
\alias{plot.FIPS_simulation}
\title{plot.FIPS_simulation}
\usage{
\method{plot}{FIPS_simulation}(.FIPS_simulation, from = NULL,
  to = NULL, plot_stat = NULL, fatigue_CIs = FALSE,
  observed = NULL, observed_y = NULL, ...)
}
\arguments{
\item{.FIPS_simulation}{A valid .FIPS_simulation series that has been simulated}

\item{from}{The starting datetime to be plotted}

\item{to}{The ending datetime to be plotted}

\item{plot_stat}{Which variables to plot}

\item{fatigue_CIs}{A logical indicating whether uncertainty intervals on fatigue should be plotted}

\item{observed}{A data frame with any observed sleepiness ratings or objective indicators to plot against predictions}

\item{observed_y}{The name of the observed sleepiness ratings in the observed data frame}
}
\description{
S3 plot method for FIPS_simulation
}
