% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIPS_simulation.R
\name{print.FIPS_simulation}
\alias{print.FIPS_simulation}
\title{print.FIPS_simulation}
\usage{
\method{print}{FIPS_simulation}(x, ...)
}
\arguments{
\item{x}{A FIPS simulation object}

\item{...}{further arguments passed to print}
}
\description{
print.FIPS_simulation
}
