% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIPS_formula_interface.R
\name{process_bmm_formula}
\alias{process_bmm_formula}
\title{process_bmm_formula}
\usage{
process_bmm_formula(.FIPS_sim, model_formula, pvec)
}
\arguments{
\item{.FIPS_sim}{A FIPS_Simulation object}

\item{model_formula}{A formula describing how the time-varying processes predictors should be calculated for the predicted output.}

\item{pvec}{A required pvec argument for the .FIPS_sim}
}
\value{
Minimal dataframe with only requird columns for computation
}
\description{
The pvec is required to ensure it is contained in the environment for expression evaluation.
}
