% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_threeprocessmodel.R
\docType{data}
\name{pvec.threeprocess}
\alias{pvec.threeprocess}
\title{pvec.threeprocess
Here for compatability - will likely remove soon}
\format{
An object of class \code{numeric} of length 15.
}
\usage{
pvec.threeprocess
}
\description{
pvec.threeprocess
Here for compatability - will likely remove soon
}
\keyword{datasets}
