% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sleeptimesformat.R
\name{round_times}
\alias{round_times}
\title{Round times by column}
\usage{
round_times(.stdata, colname, round_by = 5)
}
\arguments{
\item{.stdata}{The sleeptimes dataframe}

\item{colname}{the column required to be rounded}

\item{round_by}{Amount (in minutes) to round sleep times to}
}
\value{
The sleep dataframe with all sleep.start and sleep.end rounded to X minute interval
}
\description{
Round times by column
}
