% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIPS_simulation.R
\name{summary.FIPS_simulation}
\alias{summary.FIPS_simulation}
\title{summary.FIPS_simulation}
\usage{
\method{summary}{FIPS_simulation}(object, ...)
}
\arguments{
\item{object}{an object for which a summary is desired.}

\item{...}{additional arguments affecting the summary produced}
}
\description{
summary.FIPS_simulation
}
