% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile_plot.R
\name{tileplot_dataprep_FIPS_sim}
\alias{tileplot_dataprep_FIPS_sim}
\title{Generate a heat-tile plot}
\usage{
tileplot_dataprep_FIPS_sim(
  FIPS_sim,
  pred_stat = NULL,
  show_sleep = F,
  output_label = "Fatigue",
  dv_range = NULL,
  dv_breaks = NULL,
  plot_palette = NA
)
}
\arguments{
\item{dv_breaks}{}
}
\value{

}
\description{
Generate a heat-tile plot
}
\note{
Currently plotly is not supported, due to the axis flipping.
}
