% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_threeprocessmodel.R
\name{TPM_breaktimefun}
\alias{TPM_breaktimefun}
\title{Break time function (i.e., bt function)}
\usage{
TPM_breaktimefun(ha, g, bl, ss)
}
\arguments{
\item{ha}{High asymptote (typically = 14.3)}

\item{g}{Rate of recovery (typically about -0.3813)}

\item{bl}{Break point (typically 12.2)}

\item{ss}{S upon falling asleep}
}
\value{
breaktime
}
\description{
Break time function (i.e., bt function)
}
