% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_threeprocessmodel.R
\name{TPM_Cfun}
\alias{TPM_Cfun}
\title{TPM C Function (24-hour circadian process)}
\usage{
TPM_Cfun(Cm, Ca, p, tod)
}
\arguments{
\item{Cm}{Mesor of C process (typically = 0)}

\item{Ca}{Amplitude of C process (typically = 2.5)}

\item{p}{Default C phase (i.e., time of peak typically 16.8)}

\item{tod}{Time of day (in decimal hours)}
}
\value{
C
}
\description{
Calculates 24-hour circadian process.
}
