% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_threeprocessmodel.R
\name{TPM_check_pvec}
\alias{TPM_check_pvec}
\title{Check Three Process Parameter Vector}
\usage{
TPM_check_pvec(pvec)
}
\arguments{
\item{pvec}{The pvec to check contains all required three process model parameters}
}
\value{
logical
}
\description{
Checks the pvec contains required parameters.
}
