% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_threeprocessmodel.R
\name{TPM_get_KSS_vector}
\alias{TPM_get_KSS_vector}
\title{TPM_add_KSS}
\usage{
TPM_get_KSS_vector(.FIPS_sim)
}
\arguments{
\item{.FIPS_sim}{a FIPS_simulation on the TPM}
}
\description{
Adds the KSS to a TPM FIPS_simulation. Will return error if FIPS_simulation not a TPM.
}
