% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_threeprocessmodel.R
\name{TPM_make_pvec}
\alias{TPM_make_pvec}
\title{Make Three Process Model (TPM) Parameter Vector}
\usage{
TPM_make_pvec(
  la = 2.4,
  ha = 14.3,
  d = -0.0353,
  g = log((14.3 - 14)/(14.3 - 7.96))/8,
  bl = 12.2,
  Cm = 0,
  Ca = 2.5,
  p = 16.8,
  Um = -0.5,
  Ua = 0.5,
  Wc = -5.72,
  Wd = -1.51,
  S0 = 7.96,
  KSS_intercept = 10.6,
  KSS_beta = -0.6
)
}
\arguments{
\item{la}{Low asymptote. Minimum alertness allowed by homeostatic process (S)}

\item{ha}{High asymptote. Maximum alertness allowed by homeostatic process (S)}

\item{d}{Alertness decay rate. Rate at which alterness decays when awake (homestatic process)}

\item{g}{Fatigue recovery rate per unit time. Rate at which alertness recovers when asleep}

\item{bl}{Alertness level that breaks Sprime function. The alertness level at which low pressure sleep kicks in}

\item{Cm}{Mesor of C process (average level) for 24-hour circadian (C) process}

\item{Ca}{Amplitude of C process (extent to which peaks deviate from average level) for 24-hour circadian (C) process}

\item{p}{Default C process phase (i.e., peak) for 24-hour circadian (C) process}

\item{Um}{Mesor of U process (average level) for 12-hour circadian (U) process (dip in afternoon)}

\item{Ua}{Amplitude of U process process (extent to which peaks deviate from average level) for 12-hour circadian (U) process (dip in afternoon)}

\item{Wc}{Initial reduction in alertness for sleep inertia (W) process}

\item{Wd}{Recovery rate for sleep inertia (W) process}

\item{S0}{Initial value of homeostatic process (S)}

\item{KSS_intercept}{KSS transformation intercept}

\item{KSS_beta}{KSS transformation beta}
}
\value{
parameter vector
}
\description{
Creates and checks a TPM parameter vector. No arguments returns default settings from
Ingre, M., Van Leeuwen, W., Klemets, T., Ullvetter, C., Hough, S., Kecklund, G., Karlsson, D., & Åkerstedt, T. (2014). Validating and Extending the Three Process Model of Alertness in Airline Operations. \emph{PLoS ONE}, \emph{9}(10), e108679. \url{https://doi.org/10.1371/journal.pone.0108679}
}
