% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_threeprocessmodel.R
\name{TPM_Sfun}
\alias{TPM_Sfun}
\title{TPM Wake S Function}
\usage{
TPM_Sfun(la, d, sw, taw)
}
\arguments{
\item{la}{Lower asymptote (typically = 2.4)}

\item{d}{Decay in alertness (typically = -0.0353)}

\item{sw}{S upon waking}

\item{taw}{Time awake (hours)}
}
\value{
S
}
\description{
Calculates S process during wake
}
