% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_threeprocessmodel.R
\name{TPM_simulation_dispatch}
\alias{TPM_simulation_dispatch}
\title{TPM Simulation Dispatcher}
\usage{
TPM_simulation_dispatch(dat, pvec, model_formula)
}
\arguments{
\item{dat}{input dataframe (ensure this is in FIPS format)}

\item{pvec}{a vector of default parameters}

\item{model_formula}{A formula expression object of desired model caluclation}
}
\description{
Constructor/dispatcher for TPM model simulations.
}
