% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_threeprocessmodel.R
\name{TPM_Sp1fun}
\alias{TPM_Sp1fun}
\title{TPM_Sp1fun}
\usage{
TPM_Sp1fun(ha, g, bl, ss, tas)
}
\arguments{
\item{ha}{High asymptote (typically = 14.3)}

\item{g}{Rate of recovery (typically about -0.3813)}

\item{bl}{Break point (typically 12.2)}

\item{ss}{S upon falling asleep}

\item{tas}{Time asleep (hours)}
}
\value{
S
}
\description{
Calculates S during high pressure component of sleep before break point is reached
}
