% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_threeprocessmodel.R
\name{TPM_Sp2fun}
\alias{TPM_Sp2fun}
\title{TPM_Sp2fun}
\usage{
TPM_Sp2fun(ha, g, bl, tas, breaktime)
}
\arguments{
\item{ha}{High asymptote (typically = 14.3)}

\item{g}{Rate of recovery (typically about -0.3813)}

\item{bl}{Break point (typically 12.2)}

\item{tas}{Time asleep (hours)}

\item{breaktime}{Break time (time from sleep until breakpoint is reached)}
}
\value{
S
}
\description{
Calculates S during lower pressure component of sleep after break point is reached
}
