% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_threeprocessmodel.R
\name{TPM_Spfun}
\alias{TPM_Spfun}
\title{TPM Sleep S Function}
\usage{
TPM_Spfun(ha, g, bl, ss, tas)
}
\arguments{
\item{ha}{High asymptote (typically = 14.3)}

\item{g}{Rate of recovery (typically about -0.3813)}

\item{bl}{Break point of sleep recovery (typically 12.2)}

\item{ss}{S at falling asleep}

\item{tas}{Time asleep (hours)}
}
\value{
S
}
\description{
Calculates S during sleep
}
