% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_threeprocessmodel.R
\name{TPM_Wfun}
\alias{TPM_Wfun}
\title{Sleep Inertia (W) Function}
\usage{
TPM_Wfun(Wc, Wd, taw)
}
\arguments{
\item{Wc}{Extent of alertness reduction at time of waking (typically = -5.72)}

\item{Wd}{Exponential recovery of alertness (typically = -1.51)}

\item{taw}{Time awake (hours)}
}
\value{
W
}
\description{
Calculates effect of sleep inertia on alertness
}
