% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_unifiedfatiguemodel.R
\name{unified_Cfun}
\alias{unified_Cfun}
\title{Unified Circadian Process (C)
calculates C (circadian process)}
\usage{
unified_Cfun(tod, phi, tau = 24, A = 1)
}
\arguments{
\item{tod}{Time of day (in decimal hours)}

\item{phi}{Phase at beginning of the simulation (I think this should be 0 if t = tod)}

\item{tau}{Period of C process}

\item{A}{Amplitute of process}
}
\description{
Unified Circadian Process (C)
calculates C (circadian process)
}
