% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_unifiedfatiguemodel.R
\name{unified_check_pvec}
\alias{unified_check_pvec}
\title{Check Unified Parameter Vector}
\usage{
unified_check_pvec(pvec)
}
\arguments{
\item{pvec}{The pvec to check contains all required three process model parameters}
}
\value{
logical
}
\description{
Checks the pvec contains required parameters.
}
