% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_unifiedfatiguemodel.R
\name{unified_Lfun}
\alias{unified_Lfun}
\title{Sleep Debt Penalty (L) Function during Wake
Calculates L / Sleep Debt Process during wake}
\usage{
unified_Lfun(l_at_wake, taw, tau_la, U0)
}
\arguments{
\item{l_at_wake}{Lower asymptote at wake onset}

\item{taw}{Time awake}

\item{tau_la}{Rate of change in lower asymptote}

\item{U0}{Upper asymptote}
}
\description{
Sleep Debt Penalty (L) Function during Wake
Calculates L / Sleep Debt Process during wake
}
