% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_unifiedfatiguemodel.R
\name{unified_Lpfun}
\alias{unified_Lpfun}
\title{Sleep Debt Penalty (L) Function during Sleep
calculates L during sleep}
\usage{
unified_Lpfun(ls, tas, tau_la, U0)
}
\arguments{
\item{ls}{Lower asymptote upon falling asleep}

\item{tas}{Time asleep}

\item{tau_la}{Rate of change in lower asymptote}

\item{U0}{Upper asymptote}
}
\description{
Sleep Debt Penalty (L) Function during Sleep
calculates L during sleep
}
