% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_unifiedfatiguemodel.R
\name{unified_make_pvec}
\alias{unified_make_pvec}
\title{Make Model Default (pvec) Parameters}
\usage{
unified_make_pvec(U0 = 24.12, L0 = 0, S0 = 0, phi = 2.02,
  kappa = 4.13, tau_s = 1, tau_w = 40, tau_la = 4.06 * 24,
  sigma = 1, wc = 1.14, wd = -0.46)
}
\arguments{
\item{U0}{Upper asymptote (defaults to = 24.12)}

\item{L0}{Lower asymptote(defaults to = 0,   # (0.88 * 3 - 2)*1.74)}

\item{S0}{Initial starting point of S process (defaults to = 0,   # 1.11 + (1.74-1.11)*0.64)}

\item{phi}{Phase at beginning of the simulation (defaults to = 2.02)}

\item{kappa}{Relative influence of C process (defaults to = 4.13)}

\item{tau_s}{Controls rate of decay in S during sleep (defaults to = 1)}

\item{tau_w}{Controls rate of rise in S during wake (defaults to = 40)}

\item{tau_la}{Rate of change in lower asymptote (defaults to = 4.06*24)
I don't think we have any particular reason to claim sigma is Bayesian error.}

\item{sigma}{Bayesian error - ignore unless you have error calculations (defaults to = 1)}

\item{wc}{Sleep inertia: extent of alertness reduction at time of waking (typically = -5.72) (defaults to = 1.14)}

\item{wd}{Sleep inertia: exponential recovery of alertness (typically = -1.51) (defaults to = -0.4)}
}
\description{
The default unified model parameters from:
Ramakrishnan, S., Wesensten, N. J., Balkin, T. J., \& Reifman, J.
(2016). A Unified Model of Performance: Validation of its Predictions
across Different Sleep/Wake Schedules. \emph{Sleep}, \emph{39}(1),
249--262. \url{https://doi.org/10.5665/sleep.5358}
}
