% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_unifiedfatiguemodel.R
\docType{data}
\name{unified_pvec}
\alias{unified_pvec}
\title{Unified Model Default (pvec) Parameters}
\format{An object of class \code{numeric} of length 11.}
\usage{
unified_pvec
}
\description{
The default unified model parameters from:
}
\details{
Ramakrishnan, S., Wesensten, N. J., Balkin, T. J., \& Reifman, J.
(2016). A Unified Model of Performance: Validation of its Predictions
across Different Sleep/Wake Schedules. \emph{Sleep}, \emph{39}(1),
249--262. \url{https://doi.org/10.5665/sleep.5358}
}
\keyword{datasets}
