% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_unifiedfatiguemodel.R
\name{unified_Sfun}
\alias{unified_Sfun}
\title{S Wake (Unified)
Calculates S during wake}
\usage{
unified_Sfun(s_at_wake, taw, tau_w, U0)
}
\arguments{
\item{s_at_wake}{S upon waking}

\item{taw}{Time awake}

\item{tau_w}{Controls rate of rise in S during wake}

\item{U0}{Upper asymptote}
}
\description{
S Wake (Unified)
Calculates S during wake
}
