% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_unifiedfatiguemodel.R
\name{unified_simulate}
\alias{unified_simulate}
\title{Simulate: Unified Model}
\usage{
unified_simulate(pvec, dat)
}
\arguments{
\item{pvec}{a vector of default parameters, see [unified_pvec]}

\item{dat}{input dataframe (ensure this is a FIPS_df)}
}
\value{
simulated dataset complete
}
\description{
Runs a full simulation of the 'Unified Model'.
}
\section{References}{


Rajdev, P., Thorsley, D., Rajaraman, S., Rupp, T. L., Wesensten, N. J.,
Balkin, T. J., \& Reifman, J. (2013). A unified mathematical model to
quantify performance impairment for both chronic sleep restriction and
total sleep deprivation. \emph{Journal of Theoretical Biology},
\emph{331}, 66--77. \url{https://doi.org/10.1016/j.jtbi.2013.04.013}

Ramakrishnan, S., Wesensten, N. J., Balkin, T. J., \& Reifman, J.
(2016). A Unified Model of Performance: Validation of its Predictions
across Different Sleep/Wake Schedules. \emph{Sleep}, \emph{39}(1),
249--262. \url{https://doi.org/10.5665/sleep.5358}
}

\seealso{
unified_make_pvec
}
