% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_unifiedfatiguemodel.R
\name{unified_simulation_dispatch}
\alias{unified_simulation_dispatch}
\title{Unified Simulation Dispatcher}
\usage{
unified_simulation_dispatch(dat, pvec, model_formula)
}
\arguments{
\item{dat}{input dataframe (ensure this is a FIPS_df)}

\item{pvec}{a vector of default parameters, see [unified_pvec]}

\item{model_formula}{A formula expression object of desire model caluclation}
}
\description{
Constructor/dispatcher for Unified model simulations.
}
