% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_unifiedfatiguemodel.R
\name{unified_Spfun}
\alias{unified_Spfun}
\title{S Sleep (Unified)
Calculates S during sleep}
\usage{
unified_Spfun(ss, tas, tau_s, U0, tau_la, ls)
}
\arguments{
\item{ss}{S upon falling asleep}

\item{tas}{Time asleep}

\item{tau_s}{Controls rate of decay in S during sleep}

\item{U0}{Upper asymptote}

\item{tau_la}{Rate of change in lower asymptote}

\item{ls}{Lower asymptote at sleep onset}
}
\description{
S Sleep (Unified)
Calculates S during sleep
}
