% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_unifiedfatiguemodel.R
\name{unified_Wfun}
\alias{unified_Wfun}
\title{Sleep inertia function (direct 3PM import)
Caclulates effect of sleep intertia on alterness}
\usage{
unified_Wfun(taw, wc, wd)
}
\arguments{
\item{taw}{Time awake}

\item{wc}{Extent of alertness reduction at time of waking (typically = -5.72)}

\item{wd}{Rate of recovery of alterness (typically = -1.51)}
}
\description{
Sleep inertia function (direct 3PM import)
Caclulates effect of sleep intertia on alterness
}
