% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{county}
\alias{county}
\title{County-level FIPS Code}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3235 rows and 4 columns.}
\source{
\url{https://www.census.gov/geo/reference/codes/cou.html}
}
\usage{
county
}
\description{
2010 FIPS Codes for Counties and County Equivalent Entities.
}
\keyword{datasets}

