% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fips}
\alias{fips}
\title{State-level FIPS Code}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 57 rows and 3 columns.}
\source{
\url{https://www.census.gov/geo/reference/ansi_statetables.html}
}
\usage{
fips
}
\description{
FIPS Codes for the 50 states, the District of Columbia, and the Outlying Areas of the United States.
}
\keyword{datasets}
