% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIPS_simulate.R
\name{FIPS_simulate}
\alias{FIPS_simulate}
\title{FIPS Simulation dispatcher}
\usage{
FIPS_simulate(FIPS_df, modeltype = NULL, pvec, formula = NULL)
}
\arguments{
\item{FIPS_df}{A valid FIPS_df series that has not already been modelled}

\item{modeltype}{String: either \code{"TPM"} (Three Process Model) or \code{"unified"}.}

\item{pvec}{Parameter vector (named list), see default pvecs for guidance.}

\item{formula}{A formula to be parsed to calculate the overall fatigue/alertness score (see details).}
}
\value{
a FIPS_simulation object
}
\description{
\code{FIPS_simulate} is used to apply a particular BMM simulation to a \code{FIPS_df}.
It will dispatch to the selected model simulation type and then return the df with
the fitted model columns added on as a \code{FIPS_simulation} object.
}
