% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_mccauley2024.R
\name{mccauley2024_make_pvec}
\alias{mccauley2024_make_pvec}
\title{Make McCauley Model Parameter Vector}
\usage{
mccauley2024_make_pvec(
  alpha_w = 0.028,
  alpha_s = 0.26,
  beta_w = 0.26,
  beta_s = 0.26,
  eta_w = 0.0126,
  Wc = 20.2,
  Tp = 24,
  mu_w = 0.466,
  mu_s = -1.5,
  phi = 21.2,
  lambda_w = 0.49,
  lambda_s = -0.49,
  xi_u = 1.09,
  xi_k = 1.09,
  xi_h = 1.09,
  p0 = 5.22,
  u0 = 38.5,
  k0 = 0.0212,
  pf = 0,
  h0 = 0,
  zeta_w = 1.31,
  zeta_s = 1.31,
  v_w = 1.37,
  v_s = 1.37,
  gamma_h = 0.71
)
}
\arguments{
\item{alpha_w}{Homeostatic dissipation rate for performance during wakefulness}

\item{alpha_s}{Homeostatic dissipation rate for performance during sleep}

\item{beta_w}{Scaling factor for \code{u} during wakefulness}

\item{beta_s}{Scaling factor for \code{u} during sleep}

\item{mu_w}{Offset of circadian process during wakefulness}

\item{mu_s}{Offset of circadian process during sleep}

\item{phi}{Phase position of circadian process}

\item{p0}{Initial value for the performance process}

\item{u0}{Initial value for the homeostatic process}

\item{k0}{Initial value for the circadian amplitude}

\item{kappa}{Build-up rate for \code{u} during wakefulness}

\item{lambda}{Rate constant for modulation of circadian amplitude}

\item{A}{Asymptotic amplitude of circadian modulation}
}
\description{
Creates a parameter vector with defaults drawn from the published model. The
values provided here are placeholders that approximate those reported in the
paper and can be altered as required.
}
