% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{state}
\alias{state}
\title{State-level FIPS Code}
\format{A data frame with three variables:
\describe{
\item{fips}{FIPS State Code}
\item{usps}{Official United States Postal Service (USPS) Code}
\item{state}{State Name}
}}
\source{
\url{https://www.census.gov/geo/reference/ansi_statetables.html}

\url{https://www.census.gov/geo/reference/codes/cou.html}
}
\usage{
state

state
}
\description{
FIPS Codes for the 50 States and the District of Columbia.

2010 FIPS Codes for Counties and County Equivalent Entities.
}
\keyword{datasets}

